+++ src/server/features/fileConfigurationManager.ts
@@ -2,11 +2,10 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-import { workspace, window, WorkspaceConfiguration, disposeAll, FormattingOptions, Uri } from 'coc.nvim'
-import { CancellationToken, Disposable } from 'vscode-languageserver-protocol'
-import { TextDocument } from 'coc.nvim'
-import Proto from '../protocol'
+import { disposeAll, FormattingOptions, TextDocument, Uri, window, workspace, WorkspaceConfiguration } from 'coc.nvim'
 import path from 'path'
+import { CancellationToken, Disposable } from 'vscode-languageserver-protocol'
+import Proto from '../protocol'
 import { ITypeScriptServiceClient } from '../typescriptService'
 import API from '../utils/api'
 import { equals } from '../utils/objects'
@@ -194,11 +193,12 @@ export default class FileConfigurationManager {
       includeCompletionsWithClassMemberSnippets: config.get<boolean>('suggest.classMemberSnippets.enabled', true),
       includeCompletionsWithObjectLiteralMethodSnippets: config.get<boolean>('suggest.objectLiteralMethodSnippets.enabled', true),
       autoImportFileExcludePatterns: this.getAutoImportFileExcludePatternsPreference(preferencesConfig, workspace.getWorkspaceFolder(uri)?.uri),
+      preferTypeOnlyAutoImports: preferencesConfig.get<boolean>('preferTypeOnlyAutoImports', false),
       useLabelDetailsInCompletionEntries: true,
       allowIncompleteCompletions: true,
       displayPartsForJSDoc: true,
       ...getInlayHintsPreferences(config),
-    }
+    } as any
     return preferences
   }