+++ history.md
@@ -1,3 +1,21 @@
+# 2.3.1
+
+- Added configuration `tsserver.socketPath` for standalone tsserver (not recommended).
+- Fix `tsserver.watchBuild` not work on windows.
+
+# 2.3.0
+
+- Added configuration `tsserver.disabledSchemes`.
+- Use in memory prefix for none file buffer, avoid server error.
+
+# 2.2.5
+
+- Add configuration `typescript.preferences.preferTypeOnlyAutoImports`.
+
+# 2.2.4
+
+- Improve file select on refactor.
+
 # 2.2.3
 
 - Remove unnecessary selection on refactor.